python_reference/requirements.txt
python-dotenv
requests
validators
openai